+++ lib/mcp/server.rb
@@ -59,10 +59,7 @@ module MCP
       @server_context = server_context
       @configuration = MCP.configuration.merge(configuration)
 
-      if @configuration.protocol_version == "2024-11-05" && @instructions
-        message = "`instructions` supported by protocol version 2025-03-26 or higher"
-        raise ArgumentError, message
-      end
+      validate!
 
       @capabilities = capabilities || default_capabilities
 
@@ -102,6 +99,8 @@ module MCP
     def define_tool(name: nil, title: nil, description: nil, input_schema: nil, annotations: nil, meta: nil, &block)
       tool = Tool.define(name:, title:, description:, input_schema:, annotations:, meta:, &block)
       @tools[tool.name_value] = tool
+
+      validate!
     end
 
     def define_prompt(name: nil, title: nil, description: nil, arguments: [], &block)
@@ -171,6 +170,25 @@ module MCP
 
     private
 
+    def validate!
+      if @configuration.protocol_version == "2024-11-05"
+        if @instructions
+          message = "`instructions` supported by protocol version 2025-03-26 or higher"
+          raise ArgumentError, message
+        end
+
+        error_tool_names = @tools.each_with_object([]) do |(tool_name, tool), error_tool_names|
+          if tool.annotations
+            error_tool_names << tool_name
+          end
+        end
+        unless error_tool_names.empty?
+          message = "Error occurred in #{error_tool_names.join(", ")}. `annotations` are supported by protocol version 2025-03-26 or higher"
+          raise ArgumentError, message
+        end
+      end
+    end
+
     def handle_request(request, method)
       handler = @handlers[method]
       unless handler
@@ -222,7 +240,7 @@ module MCP
         name:,
         title:,
         version:,
-      }
+      }.compact
     end
 
     def init(request)
@@ -235,7 +253,7 @@ module MCP
     end
 
     def list_tools(request)
-      @tools.map { |_, tool| tool.to_h }
+      @tools.values.map(&:to_h)
     end
 
     def call_tool(request)
@@ -275,7 +293,7 @@ module MCP
     end
 
     def list_prompts(request)
-      @prompts.map { |_, prompt| prompt.to_h }
+      @prompts.values.map(&:to_h)
     end
 
     def get_prompt(request)
@@ -320,10 +338,8 @@ module MCP
 
     def accepts_server_context?(method_object)
       parameters = method_object.parameters
-      accepts_server_context = parameters.any? { |_type, name| name == :server_context }
-      has_kwargs = parameters.any? { |type, _| type == :keyrest }
 
-      accepts_server_context || has_kwargs
+      parameters.any? { |type, name| type == :keyrest || name == :server_context }
     end
 
     def call_tool_with_args(tool, arguments)