+++ lib/mcp/tool.rb
@@ -15,15 +15,15 @@ module MCP
       end
 
       def to_h
-        result = {
+        {
           name: name_value,
           title: title_value,
           description: description_value,
           inputSchema: input_schema_value.to_h,
-        }
-        result[:_meta] = meta_value if meta_value
-        result[:annotations] = annotations_value.to_h if annotations_value
-        result
+          outputSchema: @output_schema_value&.to_h,
+          annotations: annotations_value&.to_h,
+          _meta: meta_value,
+        }.compact
       end
 
       def inherited(subclass)
@@ -32,6 +32,7 @@ module MCP
         subclass.instance_variable_set(:@title_value, nil)
         subclass.instance_variable_set(:@description_value, nil)
         subclass.instance_variable_set(:@input_schema_value, nil)
+        subclass.instance_variable_set(:@output_schema_value, nil)
         subclass.instance_variable_set(:@annotations_value, nil)
         subclass.instance_variable_set(:@meta_value, nil)
       end
@@ -51,6 +52,8 @@ module MCP
       def input_schema_value
         @input_schema_value || InputSchema.new
       end
+
+      attr_reader :output_schema_value
 
       def title(value = NOT_SET)
         if value == NOT_SET
@@ -80,6 +83,16 @@ module MCP
         end
       end
 
+      def output_schema(value = NOT_SET)
+        if value == NOT_SET
+          output_schema_value
+        elsif value.is_a?(Hash)
+          @output_schema_value = OutputSchema.new(value)
+        elsif value.is_a?(OutputSchema)
+          @output_schema_value = value
+        end
+      end
+      
       def meta(value = NOT_SET)
         if value == NOT_SET
           @meta_value
@@ -96,13 +109,14 @@ module MCP
         end
       end
 
-      def define(name: nil, title: nil, description: nil, input_schema: nil, meta: nil, annotations: nil, &block)
+      def define(name: nil, title: nil, description: nil, input_schema: nil, output_schema: nil, meta: nil, annotations: nil, &block)
         Class.new(self) do
           tool_name name
           title title
           description description
           input_schema input_schema
-          metadata metadata
+          meta meta
+          output_schema output_schema
           self.annotations(annotations) if annotations
           define_singleton_method(:call, &block) if block
         end