+++ test/mcp/server_test.rb
@@ -1,4 +1,3 @@
-# typed: true
 # frozen_string_literal: true
 
 require "test_helper"
@@ -39,7 +38,7 @@ module MCP
         arguments: [
           Prompt::Argument.new(name: "test_argument", description: "Test argument", required: true),
         ],
-      ) do |_|
+      ) do
         Prompt::Result.new(
           description: "Hello, world!",
           messages: [
@@ -774,7 +773,7 @@ module MCP
       assert_equal Configuration::DEFAULT_PROTOCOL_VERSION, response[:result][:protocolVersion]
     end
 
-    test "server uses default title when not configured" do
+    test "server response does not include title when not configured" do
       server = Server.new(name: "test_server")
       request = {
         jsonrpc: "2.0",
@@ -783,7 +782,7 @@ module MCP
       }
 
       response = server.handle(request)
-      assert_nil response[:result][:serverInfo][:title]
+      refute response[:result][:serverInfo].key?(:title)
     end
 
     test "server uses default version when not configured" do
@@ -823,6 +822,28 @@ module MCP
         Server.new(name: "test_server", instructions: "The server instructions.", configuration: configuration)
       end
       assert_equal("`instructions` supported by protocol version 2025-03-26 or higher", exception.message)
+    end
+
+    test "server uses annotations when configured with protocol version 2025-03-26" do
+      configuration = Configuration.new(protocol_version: "2025-03-26")
+      server = Server.new(name: "test_server", configuration: configuration)
+      server.define_tool(
+        name: "defined_tool",
+        annotations: { title: "test server" },
+      )
+      assert_equal({ destructiveHint: true, idempotentHint: false, openWorldHint: true, readOnlyHint: false, title: "test server" }, server.tools.first[1].annotations.to_h)
+    end
+
+    test "raises error if annotations are used with protocol version 2024-11-05" do
+      configuration = Configuration.new(protocol_version: "2024-11-05")
+      exception = assert_raises(ArgumentError) do
+        server = Server.new(name: "test_server", configuration: configuration)
+        server.define_tool(
+          name: "defined_tool",
+          annotations: { title: "test server" },
+        )
+      end
+      assert_equal("Error occurred in defined_tool. `annotations` are supported by protocol version 2025-03-26 or higher", exception.message)
     end
 
     test "#define_tool adds a tool to the server" do