+++ test/mcp/tool/response_test.rb
@@ -38,6 +38,19 @@ module MCP
         refute response.error?
       end
 
+      test "#initialize with content and structuredContent" do
+        content = [{
+          type: "text",
+          text: "{\"code\":401,\"message\":\"Unauthorized\"}",
+        }]
+        structured_content = { code: 401, message: "Unauthorized" }
+        response = Response.new(content, structured_content: structured_content)
+
+        assert_equal content, response.content
+        assert_equal structured_content, response.structured_content
+        refute response.error?
+      end
+
       test "#error? for a standard response" do
         response = Response.new(nil, error: false)
         refute response.error?
@@ -56,7 +69,7 @@ module MCP
         response = Response.new(content)
         actual = response.to_h
 
-        assert_equal [:content, :isError].sort, actual.keys.sort
+        assert_equal [:content, :isError], actual.keys
         assert_equal content, actual[:content]
         refute actual[:isError]
       end
@@ -68,9 +81,35 @@ module MCP
         }]
         response = Response.new(content, error: true)
         actual = response.to_h
-        assert_equal [:content, :isError].sort, actual.keys.sort
+        assert_equal [:content, :isError], actual.keys
         assert_equal content, actual[:content]
         assert actual[:isError]
+      end
+
+      test "#to_h for a standard response with content and structured content" do
+        content = [{
+          type: "text",
+          text: "{\"code\":401,\"message\":\"Unauthorized\"}",
+        }]
+        structured_content = { code: 401, message: "Unauthorized" }
+        response = Response.new(content, structured_content: structured_content)
+        actual = response.to_h
+
+        assert_equal [:content, :isError, :structuredContent], actual.keys
+        assert_equal content, actual[:content]
+        assert_equal structured_content, actual[:structuredContent]
+        refute actual[:isError]
+      end
+
+      test "#to_h for a standard response with structured content only" do
+        structured_content = { code: 401, message: "Unauthorized" }
+        response = Response.new(structured_content: structured_content)
+        actual = response.to_h
+
+        assert_equal [:isError, :structuredContent], actual.keys
+        assert_nil actual[:content]
+        assert_equal structured_content, actual[:structuredContent]
+        refute actual[:isError]
       end
     end
   end